added service worker route to save static assets
