+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "JATE",
+  "name": "client",
   "scripts": {
     "dev": "webpack-dev-server",
     "build": "webpack --mode production",
@@ -16,6 +16,7 @@
     "css-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
     "http-server": "^0.11.1",
+    "mini-css-extract-plugin": "^2.4.5",
     "style-loader": "^3.2.1",
     "webpack": "^5.51.1",
     "webpack-cli": "^4.8.0",
