+++ client/webpack.config.js
@@ -17,7 +17,7 @@ module.exports = () => {
       install: "./src/js/install.js",
     },
     output: {
-      filename: "JATE.bundle.js",
+      filename: "[name].bundle.js",
       path: path.resolve(__dirname, "dist"),
     },
     plugins: [
