+++ server/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "JATE",
+    "name": "server",
     "version": "1.0.0",
     "description": "",
     "main": "server.js",
